Patches to fix lots of GCC warnings:     - Apply 'const' keywords liberally     - Avoid casts that increase alignment requirements     - Use intptr_t for casts between int and void * Also, don't generate an error when ng_ksocket sends NGM_KSOCKET_CONNECT.

+++ src/ccp.c
@@ -648,7 +648,7 @@ CcpSetCommand(int ac, char *av[], void *arg)
 
   if (ac == 0)
     return(-1);
-  switch ((int) arg) {
+  switch ((intptr_t)arg) {
     case SET_ACCEPT:
       AcceptCommand(ac, av, &ccp->options, gConfList);
       break;
